requirements.txt
# ============================================
# requirements.txt
# ============================================
"""
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
openai-whisper==20231117
pydub==0.25.1
gTTS==2.4.0
python-dateutil==2.8.2
websockets==12.0
"""